[cpu_backend] Extract shared AVX2 helpers into avx2_internal.h
Move the AVX2 helper routines shared between the FP32 kernels and the
upcoming FP16 kernels out of avx2_impl.cpp into a new avx2_internal.h
header, so both implementations can reuse them without duplication.
Pure code movement with no behavioral change.

**Self evaluation:**
Build test: [x]Passed [ ]Failed [ ]Skipped
Run test: [x]Passed [ ]Failed [ ]Skipped
